+++ src/main/java/it/ccse/uscite/application/facade/dto/input/gestionale/ApriComitato_InDTO.java
@@ -3,6 +3,8 @@
  */
 package it.ccse.uscite.application.facade.dto.input.gestionale;
 
+import java.math.BigInteger;
+
 import it.ccse.uscite.application.facade.dto.input.Base_InDTO;
 
 /**
@@ -16,19 +18,19 @@ public class ApriComitato_InDTO extends Base_InDTO{
 	 */
 	private static final long serialVersionUID = 1L;
 	
-	private Integer idOrdineDelGiorno;
+	private BigInteger idOrdineDelGiorno;
 
 	/**
 	 * @return the idOrdineDelGiorno
 	 */
-	public Integer getIdOrdineDelGiorno() {
+	public BigInteger getIdOrdineDelGiorno() {
 		return idOrdineDelGiorno;
 	}
 
 	/**
 	 * @param idOrdineDelGiorno the idOrdineDelGiorno to set
 	 */
-	public void setIdOrdineDelGiorno(Integer idOrdineDelGiorno) {
+	public void setIdOrdineDelGiorno(BigInteger idOrdineDelGiorno) {
 		this.idOrdineDelGiorno = idOrdineDelGiorno;
 	}
